Reincorporated badge based stat boosts
