+++ include/battle_util.h
@@ -127,5 +127,6 @@ bool32 CanBattlerGetOrLoseItem(u8 battlerId, u16 itemId);
 struct Pokemon *GetIllusionMonPtr(u32 battlerId);
 void ClearIllusionMon(u32 battlerId);
 bool32 SetIllusionMon(struct Pokemon *mon, u32 battlerId);
+static bool8 ShouldGetStatBadgeBoost(u16 flagId, u8 battlerId);
 
 #endif // GUARD_BATTLE_UTIL_H
