+++ src/battle_main.c
@@ -4236,7 +4236,8 @@ u32 GetBattlerTotalSpeedStat(u8 battlerId)
     // player's badge boost
     if (!(gBattleTypeFlags & (BATTLE_TYPE_LINK | BATTLE_TYPE_x2000000 | BATTLE_TYPE_FRONTIER))
         && FlagGet(FLAG_BADGE03_GET)
-        && GetBattlerSide(battlerId) == B_SIDE_PLAYER)
+        && GetBattlerSide(battlerId) == B_SIDE_PLAYER
+        && B_BADGE_BOOST == GEN_3)
     {
         speed = (speed * 110) / 100;
     }
